apps/pipeline/pyproject.toml
[tool.poetry]
name = "real-time-sales-data-analysis"
version = "0.1.0"
description = ""
authors = ["Nguyen Hoang <hoang.pqn2123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.2.3"
numpy = "^2.2.0"
kafka-python = "^2.0.2"
seaborn = "^0.13.2"
matplotlib = "^3.9.4"
sqlalchemy = "*"
psycopg2-binary = "^2.9.10"
pyspark = "3.5.3"
boto3 = "^1.35.81"
apache-airflow = "^2.10.3"
dagster = "^1.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
